Update patch set 5
Patch Set 5:

Back to spaces.

I can add back cmdline as fallback - but that would be half-baked,
not doing the complete thing.

Is there more than one DS variant on any devices?
If not, it's possible to get fallback method work equally well (as in, setting a correct model number).

Patch-set: 5