Update patch set 1
Patch Set 1: Code-Review-2

Let's not fork all the things to yukon.
It's not special, and there's no reason it can't use stuff from common. (like extract_elf_ramdisk)

A minimal variant could be added to sony/common (ifdefs could do), which does away with everything but the warmboot check. That would be useful on 8974 devices too (with the new bootloader), if needed in the future.

Patch-set: 1