Update patch set 1
Patch Set 1: Code-Review-1

There's no good reason Yukon can't use common.
I never said that, in fact I implicitly mentioned adding a minimal variant to common which does only the warmboot check and no more.
Yukon can use that, it would be the same as this, but the code removed here would be ifdefd instead.
In fact, we even considered moving the keycheck handling to the device specific classes, leaving only the helper functions in common but decided against that for the initial patch. No reason it can't be done now.

The 8974 devices don't require the bootloader for critical functionality unlike Yukon, so I see no reason to force an update now, when there's other important things to do.

Don't label me as someone resisting progress when you fail to even read my comments (not just here), misinterpret them in your own way, and when you're not wanting to work as a team, always wanting to do things your own new way instead (not the first time, so I want surprised seeing this commit)

If you want to, things can be sorted out peacefully, but if you don't, I'll get out of you way, and I expect you to do the same.

Patch-set: 1
Label: Code-Review=-1